fix class cast exception (Long can not be cast to BigInteger)

+++ ispyb-ejb/src/main/java/ispyb/server/common/services/sessions/Session3ServiceBean.java
@@ -195,7 +195,7 @@ public class Session3ServiceBean implements Session3Service, Session3ServiceLoca
 		Query query = entityManager.createNativeQuery("SELECT COUNT(*) FROM DataCollectionGroup WHERE sessionId = ?1 ")
 				.setParameter(1, sessionId);
 		try {
-			BigInteger res = (BigInteger) query.getSingleResult();
+			Long res = (Long) query.getSingleResult();
 
 			return res.intValue();
 		} catch (NoResultException e) {
